Fix sign warnings in the quickjs code

+++ shlr/qjs/src/quickjs.c
@@ -6480,7 +6480,7 @@ static void dbuf_put_leb128(DynBuf *s, uint32_t v)
 static void dbuf_put_sleb128(DynBuf *s, int32_t v1)
 {
     uint32_t v = v1;
-    dbuf_put_leb128(s, (2 * v) ^ -(v >> 31));
+    dbuf_put_leb128(s, (2 * v) ^ -(int64_t)(v >> 31));
 }
 
 static int get_leb128(uint32_t *pval, const uint8_t *buf,
@@ -6513,7 +6513,7 @@ static int get_sleb128(int32_t *pval, const uint8_t *buf,
         *pval = 0;
         return -1;
     }
-    *pval = (val >> 1) ^ -(val & 1);
+    *pval = (val >> 1) ^ -(int64_t)(val & 1);
     return ret;
 }
